+++ CHANGELOG.md
@@ -1,7 +1,6 @@
 # Release 4.9.0 (???)
 
-- Add support for PEP 705, adding `typing_extensions.ReadOnly` and the
-  `readonly=True` and `other_keys=False` arguments to `TypedDict`. Patch
+- Add support for PEP 705, adding `typing_extensions.ReadOnly`. Patch
   by Jelle Zijlstra.
 - All parameters on `NewType.__call__` are now positional-only. This means that
   the signature of `typing_extensions.NewType.__call__` now exactly matches the